hw5

+++ scrapper/src/main/java/edu/java/dto/StackoverflowResponse.java
@@ -5,7 +5,7 @@ import java.time.OffsetDateTime;
 import java.util.List;
 
 public record StackoverflowResponse(List<ItemResponse> items) {
-    public record ItemResponse(@JsonProperty("question_id") int questionId,
+    public record ItemResponse(@JsonProperty("question_id") Long questionId,
                                @JsonProperty("last_activity_date") OffsetDateTime lastActivityDate,
                                @JsonProperty("answer_count") int answerCount) {
     }
